Move logger to own module/namespace
